+++ Release Notes.md
@@ -10,5 +10,9 @@ Feel free to share your thoughts @ xgarmanboziax@gmail.com about:
 
 ** Actual:
 
+v1.1	16 August 2017:
+- Queue init function now returns queue address (enabling check for queue allocation success)
+- q_pull & q_flush alias created for earlier versions and purists respectively
+
 v1.0	14 July 2017:
 - First release